Limit maximum integer that subscriber outputs as time diff
In Rails 3, notifications send start_time and end_time. If the
nanosecond diff of these two values is larger than the maximum value of
a C int, this was throwing an error down in libusdt. As this information
is of questionable value compared with how DTrace can do time
comparisons in Rails 4, we limit this value to the maximum C int.